StopPatch returns uint8_t, 0 on success, FR_TIMEOUT=15 on timeout - sleep 5ms instead of 1ms (patch stop is not time-critical - give important processes more time)

+++ firmware/patch.h
@@ -111,7 +111,7 @@ extern uint8_t i2ctxbuf[8];
 
 void InitPatch0(void);
 int StartPatch(void);
-int StopPatch(void);
+uint8_t StopPatch(void);
 
 void start_dsp_thread(void);
